match version2 port placement to spec
The spec says that for version2 the order of fields for Inet and Inet6
is source_ip, destination_ip, source_port, destination_port. The code
instead was reading and writing source_ip, source_port, destination_ip,
destination_port. This commit changes that order, and also changes the
tests to match.

It also changes the type of the version2::ProxyAddresses::{Ipv4, Ipv6}
variants to SocketAddrs instead of (IP, Option<port>) pairs, since the
spec requires them to be there.