Refactor version2 parsing
This factors in common parts of address parsing.

There is a tweak to which errors are returned when
the buffer is both too small for its declared length
and for the minimum length of address data:
in this case, we now return an UnexpectedEof indicating
the first error, previously the second error was signaled.

Tests were updated with larger vectors so they could still
exercise InsufficientLengthSpecified error cases.